requirements.txt
requests
gspread
google-auth
beautifulsoup4
python-dotenv
openai
